feat: rework to include use vite-node PR
