+++ packages/nitro-server/src/runtime/utils/renderer/build-files.ts
@@ -21,6 +21,9 @@ const APP_ROOT_CLOSE_TAG = `</${appRootTag}>`
 const getServerEntry = () => import('#build/dist/server/server.mjs').then(r => r.default || r)
 
 // @ts-expect-error file will be produced after app build
+export const getViteNodeRunner = (): ViteNodeRunner => import.meta.dev && import('#build/dist/server/vite-node-runner.mjs').then(r => r.default || r)
+
+// @ts-expect-error file will be produced after app build
 const getClientManifest: () => Promise<Manifest> = () => import('#build/dist/server/client.manifest.mjs')
   .then(r => r.default || r)
   .then(r => typeof r === 'function' ? r() : r) as Promise<Manifest>
